Extract shared secure memory functions into common file
Refactored to eliminate code duplication between platform-specific files.

Created secure_memory_common.go (no build tags) containing:
- ScrambleBytes
- WipeString
- ClearIdentity
- SecureIdentityWrapper (and all its methods)
- ValidateIdk
- isValidIdkChar
- MaxIdkLength constant

Platform-specific files now only contain WipeBytes:
- secure_memory.go (//go:build !windows) - Unix/Linux implementation
- secure_memory_windows.go (//go:build windows) - Windows RtlSecureZeroMemory

The common file calls WipeBytes() which resolves to the correct
platform-specific implementation at compile time.

This reduces code duplication by ~131 lines and improves maintainability
by having a single source of truth for shared logic.
