+++ secure_memory_windows.go
@@ -6,8 +6,6 @@ import (
 	"runtime"
 	"syscall"
 	"unsafe"
-
-	ssp "github.com/sqrldev/server-go-ssp"
 )
 
 var (
@@ -39,137 +37,3 @@ func WipeBytes(b []byte) {
 
 	runtime.KeepAlive(b)
 }
-
-// ScrambleBytes overwrites a byte slice with pseudo-random data.
-//
-//go:noinline
-func ScrambleBytes(b []byte) {
-	if len(b) == 0 {
-		return
-	}
-
-	pattern := byte(0xAA)
-	for i := range b {
-		b[i] = pattern ^ byte(i&0xFF)
-	}
-
-	runtime.KeepAlive(b)
-}
-
-// WipeString clears a string reference and wipes a copy of its contents.
-// See secure_memory.go for full documentation and security considerations.
-//
-// This implementation avoids writing to the original string's backing memory,
-// which may reside in read-only segments and cause crashes.
-func WipeString(s *string) {
-	if s == nil {
-		return
-	}
-	if *s == "" {
-		return
-	}
-
-	// Copy string contents to a mutable byte slice
-	dataCopy := []byte(*s)
-
-	// Wipe the copy
-	WipeBytes(dataCopy)
-
-	// Clear the string reference
-	*s = ""
-
-	runtime.KeepAlive(dataCopy)
-}
-
-// ClearIdentity securely wipes all sensitive fields from a SqrlIdentity struct.
-// See secure_memory.go for full documentation.
-func ClearIdentity(identity *ssp.SqrlIdentity) {
-	if identity == nil {
-		return
-	}
-
-	WipeString(&identity.Idk)
-	WipeString(&identity.Suk)
-	WipeString(&identity.Vuk)
-	WipeString(&identity.Pidk)
-	WipeString(&identity.Rekeyed)
-
-	identity.SQRLOnly = false
-	identity.Hardlock = false
-	identity.Disabled = false
-	identity.Btn = 0
-
-	runtime.KeepAlive(identity)
-}
-
-// SecureIdentityWrapper provides RAII-style automatic cleanup for SqrlIdentity.
-type SecureIdentityWrapper struct {
-	Identity *ssp.SqrlIdentity
-	wiped    bool
-}
-
-// NewSecureIdentityWrapper creates a new wrapper around an existing identity.
-func NewSecureIdentityWrapper(identity *ssp.SqrlIdentity) *SecureIdentityWrapper {
-	return &SecureIdentityWrapper{
-		Identity: identity,
-		wiped:    false,
-	}
-}
-
-// Destroy securely wipes the identity and marks the wrapper as invalid.
-func (w *SecureIdentityWrapper) Destroy() {
-	if w == nil || w.wiped {
-		return
-	}
-
-	if w.Identity != nil {
-		ClearIdentity(w.Identity)
-		w.Identity = nil
-	}
-	w.wiped = true
-}
-
-// IsValid returns true if the wrapper still contains a valid identity.
-func (w *SecureIdentityWrapper) IsValid() bool {
-	return w != nil && !w.wiped && w.Identity != nil
-}
-
-// GetIdentity returns the wrapped identity if valid, otherwise returns nil.
-func (w *SecureIdentityWrapper) GetIdentity() *ssp.SqrlIdentity {
-	if !w.IsValid() {
-		return nil
-	}
-	return w.Identity
-}
-
-// ValidateIdk performs basic validation on an Identity Key.
-func ValidateIdk(idk string) error {
-	if idk == "" {
-		return ErrEmptyIdentityKey
-	}
-
-	if len(idk) > MaxIdkLength {
-		return ErrIdentityKeyTooLong
-	}
-
-	for _, c := range idk {
-		if !isValidIdkChar(c) {
-			return ErrInvalidIdentityKeyFormat
-		}
-	}
-
-	return nil
-}
-
-// isValidIdkChar checks if a character is valid for an Identity Key.
-func isValidIdkChar(c rune) bool {
-	return (c >= 'a' && c <= 'z') ||
-		(c >= 'A' && c <= 'Z') ||
-		(c >= '0' && c <= '9') ||
-		c == '+' || c == '/' || c == '=' || c == '-' || c == '_' || c == '.'
-}
-
-// Constants for validation
-const (
-	MaxIdkLength = 256
-)
